Added Hero section and Feature Section[WIP]
